+++ inc/utile.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Get current page depth
+ *
+ * @return integer
+ */
+function get_current_page_depth(){
+	global $wp_query;
+
+	$object = $wp_query->get_queried_object();
+	$parent_id  = $object->post_parent;
+	$depth = 0;
+	while($parent_id > 0){
+		$page = get_page($parent_id);
+		$parent_id = $page->post_parent;
+		$depth++;
+	}
+
+ 	return $depth;
+}