+++ CHANGELOG.md
@@ -20,6 +20,9 @@ plus de détails sur http://semver.org/lang/fr/
 * module `header` :
   * ajout du lien de changement de langue (avec gestion des erreurs)
 * module `toc` ajouté
+  * support du sommaire de la page actuelle (basé sur les composants `title`)
+  * support des pages parentes et soeurs
+  * utilisation d'un `walker` spécifique
 * page `404` ajoutée
 * I18n : plus de chaînes du thème traduisibles
 * groupes de champs offres d'emploi :